+++ src/mock/comment.js
@@ -1,4 +1,4 @@
-import {getRandomInteger} from "../utils";
+import {getRandomInteger} from "../utils/common";
 
 const generateEmoji = () => {
   const emojiSources = [
