+++ src/mock/filter.js
@@ -1,3 +1,4 @@
+
 const filmToFilterMap = (films) => {
   const all = films.length;
   let watchlist = 0;
