+++ src/utils/render.js
@@ -1,0 +1,39 @@
+import Abstract from "../view/abstract";
+
+export const RenderPosition = {
+  AFTERBEGIN: `afterbegin`,
+  BEFOREEND: `beforeend`,
+};
+
+
+export const render = (container, child, position) => {
+  if (container instanceof Abstract) {
+    container = container.element;
+  }
+  if (child instanceof Abstract) {
+    child = child.element;
+  }
+
+  switch (position) {
+    case RenderPosition.AFTERBEGIN:
+      container.prepend(child);
+      break;
+    case RenderPosition.BEFOREEND:
+      container.append(child);
+      break;
+  }
+};
+
+export const renderTemplate = (container, template, place) => {
+  if (container instanceof Abstract) {
+    container = container.element;
+  }
+
+  container.insertAdjacentHTML(place, template);
+};
+
+export const createElement = (template) => {
+  const newElement = document.createElement(`div`);
+  newElement.innerHTML = template;
+  return newElement.firstChild;
+};
